+++ cpp/Engine/Platform/Glfw/GlfwContext.h
@@ -27,7 +27,7 @@ namespace Engine {
 		void setTitle(const char* title);
 		/*void setCursorPos(glm::u16vec2 cursorPos);
 		void getCursorPos(glm::u16vec2& cursorPos);*/
-		void getScreenSize(glm::u16vec2& screenSize);
+		void getScreenSize(glm::ivec2& screenSize);
 		//void setCursorVisible(bool visible);
 		/*static void key_callback(GLFWwindow* window, int key, int scancode, int action, int mods);
 		static void mouseButtonCallback(GLFWwindow* window, int button, int action, int mods);
